refactor:     - namespace definition     - delete unnecessary using
