+++ TagsCloudTest/CircularCloudLayouterTests.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
-using TagsCloud;
 using TagsCloud.Extensions;
 using TagsCloud.Layouter;
