+++ TagsCloudTest/ConvertersTests.cs
@@ -1,7 +1,4 @@
 ﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using TagsCloud.TextProcessing.Converters;
 
 namespace TagsCloudTest
